requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
python_a2a
PyPDF2==3.0.1
python-docx==1.1.0
openai>=1.0.0